refactor: move code to more suitable place
