+++ crates/marine-test-macro-impl/src/marine_test/module_generator/methods_generator_utils.rs
@@ -42,24 +42,7 @@ pub(super) fn generate_module_method(
     let output_type = generate_output_type(&signature.outputs, records)?;
     let mcall = generate_marine_call(module_name, cp_setting, &signature, records)?;
 
-    let (cp, func_name) = match cp_setting {
-        CallParametersSettings::Default => {
-            let func_name = new_ident(&signature.name)?;
-            (TokenStream::new(), func_name)
-        }
-        CallParametersSettings::UserDefined => {
-            let maybe_comma = if signature.arguments.is_empty() {
-                TokenStream::new()
-            } else {
-                quote! { , }
-            };
-
-            let cp = quote! { #maybe_comma cp: marine_rs_sdk_test::CallParameters };
-            let func_name = format!("{}_cp", signature.name);
-            let func_name = new_ident(&func_name)?;
-            (cp, func_name)
-        }
-    };
+    let (cp, func_name) = generate_call_parameters(&cp_setting, signature)?;
 
     let module_method = quote! {
         pub fn #func_name(&mut self, #(#arguments),* #cp) #output_type {
@@ -80,24 +63,7 @@ pub(super) fn generate_module_method_forward(
     let output_type = generate_output_type(&signature.outputs, records)?;
     let mcall = generate_forward_call(module_name, cp_setting, &signature)?;
 
-    let (cp, func_name) = match cp_setting {
-        CallParametersSettings::Default => {
-            let func_name = new_ident(&signature.name)?;
-            (TokenStream::new(), func_name)
-        }
-        CallParametersSettings::UserDefined => {
-            let maybe_comma = if signature.arguments.is_empty() {
-                TokenStream::new()
-            } else {
-                quote! { , }
-            };
-
-            let cp = quote! { #maybe_comma cp: marine_rs_sdk_test::CallParameters };
-            let func_name = format!("{}_cp", signature.name);
-            let func_name = new_ident(&func_name)?;
-            (cp, func_name)
-        }
-    };
+    let (cp, func_name) = generate_call_parameters(&cp_setting, signature)?;
 
     let module_method = quote! {
         pub fn #func_name(&mut self, #(#arguments),* #cp) #output_type {
@@ -255,5 +221,29 @@ fn get_output_type(output_types: &[IType]) -> TResult<Option<&IType>> {
         0 => Ok(None),
         1 => Ok(Some(&output_types[0])),
         _ => Err(ManyFnOutputsUnsupported),
+    }
+}
+
+fn generate_call_parameters(
+    cp_setting: &CallParametersSettings,
+    signature: &IFunctionSignature,
+) -> TResult<(TokenStream, syn::Ident)> {
+    match cp_setting {
+        CallParametersSettings::Default => {
+            let func_name = new_ident(&signature.name)?;
+            Ok((TokenStream::new(), func_name))
+        }
+        CallParametersSettings::UserDefined => {
+            let maybe_comma = if signature.arguments.is_empty() {
+                TokenStream::new()
+            } else {
+                quote! { , }
+            };
+
+            let cp = quote! { #maybe_comma cp: marine_rs_sdk_test::CallParameters };
+            let func_name = format!("{}_cp", signature.name);
+            let func_name = new_ident(&func_name)?;
+            Ok((cp, func_name))
+        }
     }
 }
